refactor: page router 마이그레이션
