+++ src/features/posts/ui/PostListItem.tsx
@@ -1,15 +1,14 @@
 import Link from "next/link";
 import { findColorByLabel } from "@/entities/category/utils/findColorByLabel";
-import { toPlainText } from "@/entities/mdx/utils/toPlainText";
-import { Post } from "@/features/posts/model";
+import { PostMetaData } from "@/features/posts/model";
 import { parseDate } from "@/shared/lib/dayjs";
 
-export const PostListItem = ({ slug, title, content, category, createdAt }: Post) => {
+export const PostListItem = ({ slug, title, description, category, createdAt }: PostMetaData) => {
     return (
         <Link href={`/posts/${slug}`}>
             <div className="border-b-[1px] border-[#3d444d] py-4">
                 <h1 className="pb-2 text-xl font-bold text-[#3f87e5] hover:underline">{title}</h1>
-                <p className="line-clamp-2 text-sm text-[#9198a1]">{toPlainText(content)}</p>
+                <p className="line-clamp-2 text-sm text-[#9198a1]">{description}</p>
                 <div className="flex items-center justify-between">
                     <div className="flex items-center py-2">
                         <div
