+++ src/widgets/Section/Section.tsx
